third_party/blink/web_tests/http/tests/devtools/sources/debugger-ui/script-formatter-breakpoints-3-expected.txt

Tests the script formatting is working fine with breakpoints.


Running: testSetup

Running: testBreakpointsSetInFormattedAndRemoveInOriginalSource
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed
unformatted3.js:formatted:4checked    var c = 3;
Breakpoint sidebar pane while paused in raw
unformatted3.js:3checked breakpoint hit    var a=0;var b=1;var c=3;var d=4;var e=5;
Script execution resumed.
